requirements.txt
dnspython==1.15.0
python-hosts==0.3.7
